Move new contracts to `FoodDiary.Contracts.Products`
